+++ package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "bootstrap": "^5.1.3",
     "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react-dom": "^17.0.0",
+    "react-router-dom": "^5.3.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^1.0.0",
