+++ src/components/NavBar.jsx
@@ -1,18 +1,18 @@
 export const NavBar = () => {
   return (
-    <header className="d-flex justify-content-center py-3">
+    <header className="d-flex justify-content-center py-3 navbar-dark bg-dark">
       <ul className="nav nav-pills">
-        <li className="nav-item">
+        <li className="nav-item me-5">
           <a href="#" className="nav-link active" aria-current="page">
             Home
           </a>
         </li>
-        <li className="nav-item">
+        <li className="nav-item me-5">
           <a href="#" className="nav-link">
             All Pokemons
           </a>
         </li>
-        <li className="nav-item">
+        <li className="nav-item me-5">
           <a href="#" className="nav-link">
             FAQs
           </a>
