+++ src/components/PokemonNotFound.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const PokemonNotFound = () => {
+  return <div>Sorry, your Pokemon couldn't be found :c</div>;
+};
