+++ src/pages/ErrorPage.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const ErrorPage = () => {
+  return <div>404 Error lol</div>;
+};
