+++ src/pages/HomePage.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Pokedex } from "../components/Pokedex";
+
+export const HomePage = () => {
+  return (
+    <div>
+      <Pokedex />
+    </div>
+  );
+};
