+++ src/styles.css
@@ -1,0 +1,3 @@
+.white-text {
+  color: white;
+}
